Calculating the Minor Allele Frequency.

We calculate the MAF from a reference reads matrix and an alternative reads matrix. This function is intended to be used with the mitochondrial genome and not with other somatic mutations.

Usage

CalculateAlleleFrequency(reference_reads, alternative_reads, pseudo_count = 0)

Arguments

reference_reads

Reference reads matrix.

alternative_reads

List of matrices for the alternative reads.

pseudo_count

= What is the pseudo count you want to add to the reference_reads matrix. Default = 0