Determining the clonal diversity.

This function determines the clonal diversity. The clones are defined using the function ClonalDefinition. It is not required to use clones, but any column in the column data of the object can be used.

Usage

ClonalDiversity(
  se,
  grouping = "Clones",
  cells = NULL,
  diversity_measure = "EffectiveSpecies",
  base = exp(1),
  verbose = TRUE
)

Arguments

se

SummarizedExperiment object.

grouping

The meta data column used.

cells

Which cells should be used? NULL uses all cells.

diversity_measure

What diversity measure should be calculated? Can be EffectiveSpecies, shannon, simpson, invsimpson

base

The base for the diversity

verbose

Should the function be verbose? Default = TRUE