HeatmapVoi

We plot a heatmap of a set of Variants Of Interest using the Variant Allele Frequency values of a SummarizedExperiment object.

Usage

HeatmapVoi(
  SE,
  voi,
  annotation_trait = NULL,
  column_title = NULL,
  remove_empty_cells = FALSE
)

Arguments

SE

SummarizedExperiment object.

voi

Variants Of Interest.

annotation_trait

Cell Annotation at the bottom of the heat map.

column_title

The title of the heat map. Default = NULL

remove_empty_cells

Should cells that have a fraction of 0 for all variants be removed? Default = FALSE