LoadingMAEGATK_typewise

We load the MAEGATK output and transform it to be compatible with the VarTrix output. The input file is a specifically formated csv file with all the necessary information to run the analysis. Note that the source column in the input file needs to be mgaetk or mgatk for this function. This is case insensitive. If you want to only load a single sample without the use of an input file, you have to set the following variables.

samples_path

barcodes_path

patient

samples_file = NULL

Usage

LoadingMAEGATK_typewise(
  samples_file,
  samples_path = NULL,
  patient,
  patient_column = "patient",
  type_use = "scRNAseq_MT",
  chromosome_prefix = "chrM",
  min_cells = 2,
  barcodes_path = NULL,
  cellbarcode_length = 18,
  verbose = TRUE
)

Arguments

samples_file

Path to the csv file with the samples to be loaded.

samples_path

Path to the input folder.

patient

The patient you want to load.

patient_column

The column that contains the patient information. Use merge, if all samples should be merged.

type_use

The type of input. Only rows that have the specified type will be loaded.

chromosome_prefix

The prefix you want use. Default: "chrM"

min_cells

The minimum number of cells with coverage for a variant. Variants with coverage in less than this amount of cells are removed. Default = 2

barcodes_path

Path to the barcodes file tsv. Default = NULL

cellbarcode_length

The length of the cell barcode. This should be the length of the actual barcode plus two for the suffix (-1). Default = 18

verbose

Should the function be verbose? Default = TRUE