LoadingRawMatrix_typewise

A function to load a dense matrix of genotyping information and reformat it. Each matrix should have a column with the cell barcodes, one with the reference reads and one with the alternative reads. The type in the design matrix is used as the variant name.

Usage

LoadingRawMatrix_typewise(
  samples_file,
  patient,
  patient_column = "patient",
  variant_use = NULL,
  cells_include = NULL,
  cells_exclude = NULL,
  matrix_column_separator = ",",
  matrix_header = TRUE,
  cellbarcode_index = 1,
  ref_index = 2,
  alt_index = 3,
  cellbarcode_length = 18,
  verbose = TRUE
)

Arguments

samples_file

Path to the csv file with the samples to be loaded.

patient

The patient you want to load.

patient_column

The column that contains the patient information.

variant_use

The variant of the respective matrices.

cells_include

A vector of cell barcodes. Only these cells will be retained.

cells_exclude

A vector of cell barcodes. These cells will be removed from the output.

cellbarcode_length

The length of the cell barcode. This should be the length of the actual barcode plus two for the suffix (-1). Default = 18

verbose

Should the function be verbose? Default = TRUE

samples_path

Path to the input folder. Must include a barcodes file.