Loading VarTrix results for the down stream analysis.

This function loads the VarTrix results and outputs a list of two SummarizedExperiments objects, one for the somatic variants and on accompanying for the mitochondrial variants.

Usage

LoadingVarTrix(samples_file, vcf_path, vcf_path_MT, patient)

Arguments

samples_file

Path to the input file. One sample per row. It has to contain the following columns:

patient: The patient of the sample. A patient can have multiple samples (treated/untreated). All samples are merged per patient.

sample: The ID of the sample.

resource: Is this a VarTrix or MAEGATK input? Must be either VarTrix or MAEGATK.

type: What type of data? Must be one of the following: scRNAseq_Somatic, scRNAseq_MT, Amplicon_Somatic, Amplicon_MT.

input_folder: Path to the folder, where the VarTrix output is stored.

cells: Path to the barcodes file from the CellRanger output.

vcf_path

Path to the VCF file that contains the somatic variants you are interested in.

vcf_path_MT

Path to the VCF file that contains the MT variants you are interested in.

patient

The patient you want to load.

Details

The function loads the VarTrix for one of the patients in the input file.
If a patient has multiple samples, then all samples are merged.
There are two types of input data, single cell RNA sequencing and amplicon data.
scRNAseq data is the result from something like a 10X assay, while amplicon data focuses on a specific part of the genome.
Both are the output of the CellRanger pipeline. Amplicon data has a much higher coverage in the specific area and therefore is much better for these areas. The scRNAseq data is overwritten if amplicon data is available.

The output is a list of two SummarizedExperiment objects. Each object contains two assays:

consensus_somatic: The consensus information per somatic variant and cell.
consensus_MT: The consensus information per MT variant and cell.
fraction_somatic: The variant allele frequency per somatic variant and cell.
fraction_MT: The variant allele frequency per MT variant and cell.

The consensus information is encoded in the following way:

0: No Call, no reads are covering this position.
1: Reference, only reference reads cover this position.
2: Alternative, only mutated reads cover this position.
3: Both, both alt and ref reads cover this position.

Value

The return list has two objects.

SE_Somatic

The somatic variants results.

SE_MT

The MT variants results.

References

Author

Note

See also

Examples